Moved the thread pool class
