+++ Util/src/main/java/io/deephaven/util/datastructures/ThreadSafeMaxSizePool.java
@@ -1,31 +1,38 @@
 //
 // Copyright (c) 2016-2025 Deephaven Data Labs and Patent Pending
 //
-package io.deephaven.extensions.s3;
+package io.deephaven.util.datastructures;
 
 import io.deephaven.UncheckedDeephavenException;
 import io.deephaven.base.pool.Pool;
-import io.deephaven.util.datastructures.SegmentedSoftPool;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.concurrent.Semaphore;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
 /**
- * A thread-safe pool that holds a fixed maximum number of items.
+ * A pool that
+ * <UL>
+ * <LI>holds at max <code>maxSize</code> items,
+ * <LI>starts with no items in the pool,
+ * <LI>blocks (busily) whenever the pool overflows,
+ * <LI>optionally clears the items given to it, and
+ * <LI>is thread-safe
+ * </UL>
  */
-final class ThreadSafeMaxSizePool<ELEMENT_TYPE> implements Pool<ELEMENT_TYPE> {
+public final class ThreadSafeMaxSizePool<ELEMENT_TYPE> implements Pool<ELEMENT_TYPE> {
 
     private static final int DELEGATE_POOL_SEGMENT_CAPACITY = 10;
 
     private final Semaphore semaphore;
     private final SegmentedSoftPool<ELEMENT_TYPE> delegatePool;
 
-    ThreadSafeMaxSizePool(
+    public ThreadSafeMaxSizePool(
             final int maxSize,
             @NotNull final Supplier<ELEMENT_TYPE> factory,
-            @NotNull final Consumer<ELEMENT_TYPE> clearingProcedure) {
+            @Nullable final Consumer<ELEMENT_TYPE> clearingProcedure) {
         this.semaphore = new Semaphore(maxSize);
         this.delegatePool = new SegmentedSoftPool<>(DELEGATE_POOL_SEGMENT_CAPACITY, factory, clearingProcedure);
     }
